Add gradle support to build and run edith
