+++ src/main/java/edith/command/DeleteCommand.java
@@ -1,17 +1,17 @@
-package Edith.command;
+package edith.command;
 
-import Edith.Ui;
-import Edith.Storage;
-import Edith.EdithException;
-import Edith.task.Task;
-import Edith.task.TaskList;
+import edith.Ui;
+import edith.Storage;
+import edith.EdithException;
+import edith.task.Task;
+import edith.task.TaskList;
 
 import java.io.IOException;
 
-public class MarkCommand extends Command {
+public class DeleteCommand extends Command {
     private int index;
 
-    public MarkCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
 
@@ -21,11 +21,12 @@ public class MarkCommand extends Command {
             throw new EdithException("Edith.task.Task " + index + " does not exist. Please enter a valid Edith.task number.");
         }
 
-        Task taskToMark = tasks.getTask(index);
-        taskToMark.markTaskDone();
+        Task taskToDelete = tasks.getTask(index);
+        tasks.deleteTask(index);
 
-        ui.showIndentedMessage("Alright, great job! I've marked Edith.task " + (index + 1) + " as done:");
-        ui.showIndentedMessage(taskToMark.toString());
+        ui.showIndentedMessage("Certainly. I've removed this Edith.task:");
+        ui.showIndentedMessage(taskToDelete.toString());
+        ui.showIndentedMessage("There are now " + tasks.getNumOfTasks() + " tasks in your list.");
         ui.showLineBreak();
 
         try {
