+++ src/main/java/edith/command/UnmarkCommand.java
@@ -1,17 +1,17 @@
-package Edith.command;
+package edith.command;
 
-import Edith.Ui;
-import Edith.Storage;
-import Edith.EdithException;
-import Edith.task.Task;
-import Edith.task.TaskList;
+import edith.Ui;
+import edith.Storage;
+import edith.EdithException;
+import edith.task.Task;
+import edith.task.TaskList;
 
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     private int index;
 
-    public DeleteCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
@@ -21,12 +21,11 @@ public class DeleteCommand extends Command {
             throw new EdithException("Edith.task.Task " + index + " does not exist. Please enter a valid Edith.task number.");
         }
 
-        Task taskToDelete = tasks.getTask(index);
-        tasks.deleteTask(index);
+        Task taskToUnmark = tasks.getTask(index);
+        taskToUnmark.markTaskUndone();
 
-        ui.showIndentedMessage("Certainly. I've removed this Edith.task:");
-        ui.showIndentedMessage(taskToDelete.toString());
-        ui.showIndentedMessage("There are now " + tasks.getNumOfTasks() + " tasks in your list.");
+        ui.showIndentedMessage("Sure, I've marked Edith.task " + (index + 1) + " as not done yet:");
+        ui.showIndentedMessage(taskToUnmark.toString());
         ui.showLineBreak();
 
         try {
