+++ src/main/java/edith/task/ToDo.java
@@ -1,4 +1,4 @@
-package Edith.task;
+package edith.task;
 
 public class ToDo extends Task {
     public ToDo(String task) {
